packages businesslogic, database, gui + refactor
