+++ src/code/businesslogic/ATM.java
@@ -1,4 +1,13 @@
-package code;
+package code.businesslogic;
+
+import code.database.BankDatabase;
+import code.database.Transaction;
+import code.gui.CashDispenser;
+import code.gui.Deposit;
+import code.gui.DepositSlot;
+import code.gui.Keypad;
+import code.gui.Screen;
+import code.gui.Withdrawal;
 
 // ATM.java
 // Represents an automated teller machine
