+++ src/code/database/Account.java
@@ -1,4 +1,4 @@
-package code;
+package code.database;
 
 // Account.java
 // Represents a bank account
