+++ src/code/database/Transaction.java
@@ -1,4 +1,6 @@
-package code;
+package code.database;
+
+import code.gui.Screen;
 
 // Transaction.java
 // Abstract superclass Transaction represents an ATM transaction
