+++ src/code/gui/CashDispenser.java
@@ -1,4 +1,6 @@
-package code;
+package code.gui;
+
+import code.database.Euro;
 
 // CashDispenser.java
 // Represents the cash dispenser of the ATM
@@ -7,7 +9,7 @@ public class CashDispenser
 {
    // the default initial number of bills in the cash dispenser
    private final static int INITIAL_COUNT = 500;
-   private int count; // number of $20 bills remaining
+   private int count; // number of €20 bills remaining
    
    // no-argument CashDispenser constructor initializes count to default
    public CashDispenser()
@@ -16,16 +18,16 @@ public class CashDispenser
    } // end CashDispenser constructor
 
    // simulates dispensing of specified amount of cash
-   public void dispenseCash( int amount )
+   public void dispenseCash( Euro amount )
    {
-      int billsRequired = amount / 20; // number of $20 bills required
+      int billsRequired = amount.getEuro() / 20; // number of €20 bills required
       count -= billsRequired; // update the count of bills
    } // end method dispenseCash
 
    // indicates whether cash dispenser can dispense desired amount
-   public boolean isSufficientCashAvailable( int amount )
+   public boolean isSufficientCashAvailable( Euro amount )
    {
-      int billsRequired = amount / 20; // number of $20 bills required
+      int billsRequired = amount.getEuro() / 20; // number of €20 bills required
 
       if ( count >= billsRequired  )
          return true; // enough bills available
